Refactored types and fixed json schema
