+++ src/types/utils.ts
@@ -1,0 +1,7 @@
+export class ChildProcessError extends Error {
+    stderr: string = "";
+}
+
+export type ChildProcessResult = {
+    stdout: string;
+}
